fix: Fix bug and add slight optimization
Bug allowed 'g' and 'G' to be present in string,
which would've caused a segfault under the new
optimization. Under the old optimization, having
a 'g'/'G' may have also caused a segfault.

Optimization is more of a memory optimization. Instead
of a giant 16x16 2D matrix, we condense it to an array
of size 16 with the popcount (i.e., number of 1s in the binary)
of 0 through 15.